+++ src/utilities/dom/index.ts
@@ -2,3 +2,4 @@ export { getRectFromCoordinates } from "./flyout";
 export { getShadowRoot } from "./shadowDom";
 export { enableUserSelect, disableUserSelect } from "./userSelect";
 export { findParent, findClosestScrollableContainer, findClosestPositionContainer } from "./find";
+export { triggerChangeEvent } from "./event";